+++ releasenotes.md
@@ -1,6 +1,7 @@
 ### DOOM Retro v3.5.11
 
 * Optimizations have been made to further improve the overall performance and stability of *DOOM Retro*.
+* Minor changes have been made to the character set used in the console.
 * Minor changes have been made to text that is output to the console.
 * The following changes have been made to the silhouettes of weapons in the alternate widescreen HUD:
   * Minor changes have been made to some of them.